requirements.txt
langchain
langchain-openai
langchain-core
langgraph
pydantic
python-dotenv
pdfplumber
jupyter
